+++ user-service/src/main/java/sparta/userservice/provider/email/EmailProvider.java
@@ -1,10 +1,8 @@
-package sparta.userservice.provider;
+package sparta.userservice.provider.email;
 
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMailMessage;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Component;
