+++ user-service/src/main/java/sparta/userservice/provider/jwt/JwtBlacklist.java
@@ -1,4 +1,4 @@
-package sparta.userservice.jwt;
+package sparta.userservice.provider.jwt;
 
 import org.springframework.stereotype.Component;
